Move IKleinRequest to _interfaces.py and remove _iapp.py.
