+++ tox.ini
@@ -300,9 +300,6 @@ warn_return_any          = False
 [mypy-klein._app]
 allow_untyped_defs = True
 
-[mypy-klein._iapp]
-allow_untyped_defs = True
-
 [mypy-klein._plating]
 allow_untyped_defs = True
